backend/requirements.txt
flask>=3.0
paho-mqtt>=1.6
ruamel.yaml>=0.18
